Plot All Transcripts

plot expression all transcripts for an input gene superimposed on an embedding

plot_all_transcripts(
  seu,
  features,
  embedding = "umap",
  from_gene = TRUE,
  combine = TRUE
)

Arguments

seu

A seurat object

features

gene or vector of transcripts

embedding

umap

from_gene

whether to look up transcripts for an input gene

Examples

processed_seu <- clustering_workflow(human_gene_transcript_seu)
#> Centering and scaling data matrix
#> Centering and scaling data matrix
#> PC_ 1 
#> Positive:  CNGB1, RBP3, GUCA1A, IMPG1, GUCA1B, ROM1, PDE6B, GNAT1, LINC02528, NR2E3 
#> 	   GNGT1, GUCA1C, MYL4, ARR3, SPINK4, AC023905.1, SAG, GNB1, ISOC1, VTN 
#> 	   SLC39A5, AC112206.2, CCKBR, CNGB3, AP000997.1, SLC40A1, MT1X, FABP12, PDE6A, AC118942.1 
#> Negative:  AC011270.2, IFITM3, IFITM2, CLU, SFRP2, DKK3, CCND1, SERPING1, CA14, VIM 
#> 	   HES1, IFITM1, MDK, BCHE, NKAIN4, CRABP1, PAX6, CYR61, LAMP5, RAB31 
#> 	   TRH, MYB, CYP26A1, FBLN1, PRSS35, TMEM98, PON2, CDH11, ASCL1, OAF 
#> PC_ 2 
#> Positive:  GNAT1, GNGT1, GNB1, ROM1, NR2E3, SAG, CNGB1, PDE6A, AC002076.2, YBX3 
#> 	   RHO, LINC02115, CLUL1, PDE6B, HTRA1, PPEF2, WAPL, AP000997.1, CCKBR, DUSP1 
#> 	   IRX6, GABRR3, CTSV, MRLN, PDLIM1, IMPG1, SLC24A1, ID1, RBP3, NT5E 
#> Negative:  MIR7-3HG, RBP4, ISL2, THRB, DCT, DLL3, ARR3, APOE, SPON2, NNAT 
#> 	   DCX, SYT4, SSTR2, TMSB4X, STMN2, TMSB4XP8, ANXA2, C15orf59-AS1, FILIP1L, SLC1A6 
#> 	   AL359757.1, DPP10, CLDN5, MUSTN1, RASD1, SCD, KLHL35, AC023905.1, GRIA3, FAM19A3 
#> PC_ 3 
#> Positive:  WIF1, ANGPTL1, DACT2, CFI, C1orf61, OAF, PYGL, CP, PLP1, F3 
#> 	   FKBP10, DKK3, TOX3, CREB5, TRPM3, RRH, HLA-E, SERPING1, CD44, COL9A3 
#> 	   LAMP5, RBP1, TRH, GDPD2, WWTR1, FRZB, MIR100HG, EFEMP1, SPTLC3, SELENOP 
#> Negative:  AURKB, DLGAP5, PBK, BIRC5, TPX2, CCNB2, RRM2, CDC20, CCNA2, KIF20A 
#> 	   KIF23, PIMREG, TROAP, BUB1, SPC25, HJURP, NUF2, NEIL3, ANLN, POC1A 
#> 	   CDK1, MELK, CDCA8, OIP5, KIF2C, GTSE1, ZWINT, PRC1, PLK1, CENPM 
#> PC_ 4 
#> Positive:  GUCA1C, AC023905.1, CNGB3, MYL4, GUCA1B, ARR3, GSN, GUCA1A, AC112206.2, AC103740.1 
#> 	   RAB41, TTR, TDRG1, LINC02528, SLC39A5, AC092966.1, MT1F, GLYATL1, GPX3, OPN1MW 
#> 	   OPN1MW2, FAIM, OPN1LW, OPN1MW3, IGFBP7, C9orf135, CXCL14, IMPG1, TRPM3, DMKN 
#> Negative:  TMSB4XP8, TMSB4X, STMN2, RBP1, SSTR2, DCX, NNAT, MRLN, SYT4, DCT 
#> 	   H2AFY2, TMSB15A, DLL3, RASD1, IRX6, NR2E3, GALP, GSG1, AC010247.2, SH3BGRL 
#> 	   CRABP1, CPLX3, SNCG, TMEM176A, TMEM176B, SCHIP1, SPON2, DPP10, ISOC1, FILIP1L 
#> PC_ 5 
#> Positive:  VAT1L, PIP5K1B, HTR2C, GMNC, TPD52L1, CXCL14, DMKN, AL590094.1, CPS1, S100A11 
#> 	   GMPR, HSD17B2, ANXA1, LGALS1, ALDH1A1, BLNK, ZIC4-AS1, TM4SF1, TYR, AC073349.1 
#> 	   C21orf62, AL049777.1, CA12, ADH1C, ZIC4, AIFM3, MGST2, RHOJ, TRPM3, GPNMB 
#> Negative:  WIF1, LAMP5, GUCA1C, SFRP2, AC023905.1, ANGPTL1, ARR3, CNGB3, TRH, AC011270.2 
#> 	   PRSS35, MYL4, DACT2, GUCA1B, BCHE, LINC01658, WWTR1, RAB41, COL2A1, VSX2 
#> 	   CRYM, CDH11, MLLT3, TTYH1, GUCA1A, AC112206.2, AC103740.1, EFNA1, CREB5, HES5 
#> Warning: The following arguments are not used: check_duplicates
#> 13:35:08 UMAP embedding parameters a = 0.9922 b = 1.112
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by 'spam'
#> 13:35:08 Read 883 rows and found 30 numeric columns
#> 13:35:08 Using Annoy for neighbor search, n_neighbors = 30
#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'
#> Also defined by 'spam'
#> 13:35:08 Building Annoy index with metric = cosine, n_trees = 50
#> 0%   10   20   30   40   50   60   70   80   90   100%
#> [----|----|----|----|----|----|----|----|----|----|
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> *
#> |
#> 13:35:08 Writing NN index file to temp file /tmp/RtmpLoWyKl/file3206323b8cf03
#> 13:35:08 Searching Annoy index using 1 thread, search_k = 3000
#> 13:35:08 Annoy recall = 100%
#> 13:35:09 Commencing smooth kNN distance calibration using 1 thread
#>  with target n_neighbors = 30
#> 13:35:10 Initializing from normalized Laplacian + noise (using RSpectra)
#> 13:35:10 Commencing optimization for 500 epochs, with 35966 positive edges
#> 13:35:12 Optimization finished
#> [13:35:12] Clustering Cells...
#> Computing nearest neighbor graph
#> Computing SNN
#> clustering at 0.2 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.8961
#> Number of communities: 3
#> Elapsed time: 0 seconds
#> clustering at 0.4 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.8291
#> Number of communities: 4
#> Elapsed time: 0 seconds
#> clustering at 0.6 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.7898
#> Number of communities: 6
#> Elapsed time: 0 seconds
#> clustering at 0.8 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.7542
#> Number of communities: 6
#> Elapsed time: 0 seconds
#> clustering at 1 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.7208
#> Number of communities: 7
#> Elapsed time: 0 seconds
#> clustering at 1.2 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.6892
#> Number of communities: 10
#> Elapsed time: 0 seconds
#> clustering at 1.4 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.6613
#> Number of communities: 11
#> Elapsed time: 0 seconds
#> clustering at 1.6 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.6345
#> Number of communities: 11
#> Elapsed time: 0 seconds
#> clustering at 1.8 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.6079
#> Number of communities: 11
#> Elapsed time: 0 seconds
#> clustering at 2 resolution
#> Modularity Optimizer version 1.3.0 by Ludo Waltman and Nees Jan van Eck
#> 
#> Number of nodes: 883
#> Number of edges: 37588
#> 
#> Running Louvain algorithm...
#> Maximum modularity in 10 random starts: 0.5815
#> Number of communities: 11
#> Elapsed time: 0 seconds
#> stashing presto markers for gene_snn_res.0.2
#> stashing presto markers for gene_snn_res.0.4
#> stashing presto markers for gene_snn_res.0.6
#> stashing presto markers for gene_snn_res.0.8
#> stashing presto markers for gene_snn_res.1
#> stashing presto markers for gene_snn_res.1.2
#> stashing presto markers for gene_snn_res.1.4
#> stashing presto markers for gene_snn_res.1.6
#> stashing presto markers for gene_snn_res.1.8
#> stashing presto markers for gene_snn_res.2
#> Warning: The following features are not present in the object: MCM5, PCNA, FEN1, MCM2, RRM1, UNG, GINS2, MCM6, CDCA7, PRIM1, UHRF1, MLF1IP, RFC2, RPA2, NASP, GMNN, WDR76, SLBP, UBR7, POLD3, MSH2, ATAD2, EXO1, TIPIN, DSCC1, BLM, CASP8AP2, CLSPN, POLA1, CHAF1B, BRIP1, E2F8, not searching for symbol synonyms
#> Warning: The following features are not present in the object: TMPO, CENPF, FAM64A, CKAP2, KIF11, ANP32E, TUBB4B, KIF20B, CDCA3, HN1, TTK, RANGAP1, NCAPD2, CDCA2, ECT2, HMMR, PSRC1, LBR, CKAP5, CTCF, G2E3, CBX5, not searching for symbol synonyms
#> [13:35:14] Logging Technical Details...
transcripts_to_plot <- genes_to_transcripts("RXRG")
plot_all_transcripts(processed_seu, features = transcripts_to_plot)
#> Error in FetchData(seu, features): None of the requested variables were found: