Plot Cluster Marker Genes

Plot a dot plot of n marker features grouped by cell metadata available methods are wilcoxon rank-sum test implemented in presto and specificity scores implemented in genesorteR

plot_markers(
  seu,
  metavar = "batch",
  num_markers = 5,
  selected_values = NULL,
  return_plotly = FALSE,
  marker_method = "presto",
  seurat_assay = "gene",
  hide_technical = NULL,
  unique_markers = FALSE,
  p_val_cutoff = 1,
  ...
)

Arguments

seu

a seurat object

metavar

the metadata variable from which to pick clusters

num_markers

default is 5

return_plotly

whether to return an interactive ploly plot

marker_method

either "presto" or "genesorteR"

hide_technical

whether to exclude mitochondrial or ribosomal genes

...

Examples

# interactive mode using "presto"
plot_markers(human_gene_transcript_seu, metavar = "tech", marker_method = "presto", return_plotly = TRUE)
#> Error in `Idents<-`(`*tmp*`, value = seu[[]][[metavar]]): 'value' must be a factor or vector

# static mode using "presto"
plot_markers(human_gene_transcript_seu, metavar = "tech", marker_method = "genesorteR", return_plotly = FALSE)
#> Error in `Idents<-`(`*tmp*`, value = seu[[]][[metavar]]): 'value' must be a factor or vector