Plot Annotated Complexheatmap from Seurat object

Plot Annotated Complexheatmap from Seurat object

seu_complex_heatmap(
  seu,
  features = NULL,
  group.by = "ident",
  cells = NULL,
  layer = "scale.data",
  assay = NULL,
  group.bar.height = 0.01,
  column_split = NULL,
  col_arrangement = "ward.D2",
  mm_col_dend = 30,
  embedding = "pca",
  ...
)

Arguments

features

Vector of features to plot. Features can come

group.by

Name of one or more metadata columns to annotate columns by (for example, orig.ident)

cells

Cells to retain

column_split

whether to split columns by metadat value

col_arrangement

how to arrange columns whether with a dendrogram (Ward.D2, average, etc.) or exclusively by metadata category

mm_col_dend

height of column dendrogram

...

additional arguments passed to ComplexHeatmap::Heatmap

object

A Seurat object

emedding

for column dendrogram

Examples

# plot top 50 variable genes
top_50_features <- VariableFeatures(human_gene_transcript_seu)[1:50]
seu_complex_heatmap(human_gene_transcript_seu, features = top_50_features)
#> Error in dplyr::mutate(., across(where(is.character), ~str_wrap(str_replace_all(.x,     ",", " "), 10))): ℹ In argument: `across(...)`.
#> Caused by error in `across()`:
#> ! Can't compute column `sample_id`.
#> Caused by error in `str_wrap()`:
#> ! could not find function "str_wrap"